refactor(cli): consolidate air-gapped detection into shared utility function
- Create shared detectAirGappedMode() function in utils.ts with 30-second timeout
- Both ProtobufIRGenerator and ProtobufOpenAPIGenerator now use the shared function
- Remove backup try-catch logic - trust the initial detection result
- Simplify doGenerateLocal() in both generators to use isAirGapped flag directly
