+++ packages/cli/workspace/lazy-fern-workspace/src/protobuf/ProtobufIRGenerator.ts
@@ -1,11 +1,12 @@
 import { AbsoluteFilePath, join, RelativeFilePath } from "@fern-api/fs-utils";
 import { createLoggingExecutable, runExeca } from "@fern-api/logging-execa";
 import { TaskContext } from "@fern-api/task-context";
-import { access, chmod, cp, unlink, writeFile } from "fs/promises";
+import { chmod, cp, unlink, writeFile } from "fs/promises";
 import path from "path";
 import tmp from "tmp-promise";
 
 import {
+    detectAirGappedMode,
     getProtobufYamlV1,
     PROTOBUF_EXPORT_CONFIG_V1,
     PROTOBUF_EXPORT_CONFIG_V2,
@@ -16,19 +17,6 @@ import {
     PROTOBUF_SHELL_PROXY,
     PROTOBUF_SHELL_PROXY_FILENAME
 } from "./utils";
-
-// Network error detection helper
-function isNetworkError(errorMessage: string): boolean {
-    return (
-        errorMessage.includes("server hosted at that remote is unavailable") ||
-        errorMessage.includes("failed to connect") ||
-        errorMessage.includes("network") ||
-        errorMessage.includes("ENOTFOUND") ||
-        errorMessage.includes("ETIMEDOUT") ||
-        errorMessage.includes("TIMEDOUT") ||
-        errorMessage.includes("timed out")
-    );
-}
 
 export class ProtobufIRGenerator {
     private context: TaskContext;
@@ -70,7 +58,7 @@ export class ProtobufIRGenerator {
     }): Promise<AbsoluteFilePath> {
         // Detect air-gapped mode once at the start if we have dependencies
         if (deps.length > 0 && this.isAirGapped === undefined) {
-            await this.detectAirGappedMode(absoluteFilepathToProtobufRoot);
+            this.isAirGapped = await detectAirGappedMode(absoluteFilepathToProtobufRoot, this.context.logger);
         }
 
         const protobufGeneratorConfigPath = await this.setupProtobufGeneratorConfig({
@@ -81,76 +69,6 @@ export class ProtobufIRGenerator {
             cwd: protobufGeneratorConfigPath,
             deps
         });
-    }
-
-    /**
-     * Detect if we're in an air-gapped environment by trying buf dep update once.
-     * This sets the isAirGapped flag which is used by subsequent operations.
-     */
-    private async detectAirGappedMode(absoluteFilepathToProtobufRoot: AbsoluteFilePath): Promise<void> {
-        const bufLockPath = join(absoluteFilepathToProtobufRoot, RelativeFilePath.of("buf.lock"));
-
-        // Check if buf.lock exists (required for air-gapped mode)
-        let bufLockExists = false;
-        try {
-            await access(bufLockPath);
-            bufLockExists = true;
-            this.context.logger.debug(`Found buf.lock at: ${bufLockPath}`);
-        } catch {
-            this.context.logger.debug(`No buf.lock found at: ${bufLockPath}`);
-        }
-
-        if (!bufLockExists) {
-            // No buf.lock means we need network access - not air-gapped
-            this.isAirGapped = false;
-            return;
-        }
-
-        // Try a quick network check with buf dep update using a short timeout
-        const tmpDir = AbsoluteFilePath.of((await tmp.dir()).path);
-        try {
-            // Copy buf.yaml and buf.lock to temp dir for the test
-            const bufYamlPath = join(absoluteFilepathToProtobufRoot, RelativeFilePath.of("buf.yaml"));
-            try {
-                await cp(bufYamlPath, join(tmpDir, RelativeFilePath.of("buf.yaml")));
-                await cp(bufLockPath, join(tmpDir, RelativeFilePath.of("buf.lock")));
-            } catch {
-                // If we can't copy the files, assume not air-gapped
-                this.isAirGapped = false;
-                return;
-            }
-
-            // Try buf dep update with a short timeout (5 seconds)
-            try {
-                await runExeca(this.context.logger, "buf", ["dep", "update"], {
-                    cwd: tmpDir,
-                    stdio: "pipe",
-                    timeout: 5000 // 5 second timeout for quick detection
-                });
-                // Network works - not air-gapped
-                this.isAirGapped = false;
-                this.context.logger.debug("Network check succeeded - not in air-gapped mode");
-            } catch (error) {
-                const errorMessage = error instanceof Error ? error.message : String(error);
-                if (isNetworkError(errorMessage)) {
-                    this.isAirGapped = true;
-                    this.context.logger.debug(
-                        `Network check failed - entering air-gapped mode: ${errorMessage.substring(0, 100)}`
-                    );
-                } else {
-                    // Non-network error - assume not air-gapped
-                    this.isAirGapped = false;
-                }
-            }
-        } finally {
-            // Cleanup temp dir
-            try {
-                await unlink(join(tmpDir, RelativeFilePath.of("buf.yaml")));
-                await unlink(join(tmpDir, RelativeFilePath.of("buf.lock")));
-            } catch {
-                // Ignore cleanup errors
-            }
-        }
     }
 
     private async setupProtobufGeneratorConfig({
@@ -218,9 +136,6 @@ export class ProtobufIRGenerator {
 
         // Create a temporary buf config file to prevent conflicts
         // Try buf export with v1 first, then fall back to v2 if it fails
-        // If buf export fails due to network issues, fall back to copying files directly (backup)
-        let bufExportSucceeded = false;
-
         for (const version of ["v1", "v2"]) {
             this.context.logger.info(`Using buf export with version: ${version}`);
 
@@ -253,31 +168,13 @@ export class ProtobufIRGenerator {
                 }
 
                 await tmpBufConfigFile.cleanup();
-                bufExportSucceeded = true;
                 return;
             } catch (error) {
                 await tmpBufConfigFile.cleanup();
                 if (version === "v2") {
-                    // Backup: Check if this is a network error - if so, fall back to copying files
-                    const errorMessage = error instanceof Error ? error.message : String(error);
-                    if (isNetworkError(errorMessage)) {
-                        this.context.logger.debug(
-                            `buf export failed due to network error (backup fallback): ${errorMessage.substring(0, 200)}`
-                        );
-                        break; // Fall through to copy fallback
-                    }
                     throw error;
                 }
             }
-        }
-
-        // Backup fallback: If buf export failed due to network issues, copy files directly
-        if (!bufExportSucceeded) {
-            this.context.logger.debug("buf export failed, falling back to copying proto files directly (backup)");
-            await this.copyProtobufFilesFromRoot({
-                protobufGeneratorConfigPath,
-                absoluteFilepathToProtobufRoot
-            });
         }
     }
 
@@ -361,8 +258,6 @@ export class ProtobufIRGenerator {
             stderr: "pipe"
         });
 
-        const bufLockPath = join(cwd, RelativeFilePath.of("buf.lock"));
-
         try {
             await writeFile(bufYamlPath, configContent);
 
@@ -371,36 +266,8 @@ export class ProtobufIRGenerator {
                 if (this.isAirGapped) {
                     this.context.logger.debug("Air-gapped mode: skipping buf dep update");
                 } else {
-                    // Try buf dep update to populate the cache (needed at build time)
-                    // Backup: If it fails with a network error and buf.lock exists, continue
-                    try {
-                        await buf(["dep", "update"]);
-                    } catch (bufDepUpdateError: unknown) {
-                        const errorMessage =
-                            bufDepUpdateError instanceof Error ? bufDepUpdateError.message : String(bufDepUpdateError);
-
-                        // Check if buf.lock exists for backup fallback
-                        let bufLockExists = false;
-                        try {
-                            await access(bufLockPath);
-                            bufLockExists = true;
-                        } catch {
-                            bufLockExists = false;
-                        }
-
-                        this.context.logger.debug(
-                            `buf dep update failed. isNetworkError=${isNetworkError(errorMessage)}, bufLockExists=${bufLockExists}, error=${errorMessage.substring(0, 200)}`
-                        );
-
-                        if (isNetworkError(errorMessage) && bufLockExists) {
-                            // Backup: Air-gapped environment with pre-cached buf.lock - continue
-                            this.context.logger.debug(
-                                "buf dep update failed due to network error, but buf.lock exists. Continuing (backup)."
-                            );
-                        } else {
-                            this.context.failAndThrow(errorMessage);
-                        }
-                    }
+                    // Run buf dep update to populate the cache (needed at build time)
+                    await buf(["dep", "update"]);
                 }
             }
