+++ packages/cli/workspace/lazy-fern-workspace/src/protobuf/ProtobufOpenAPIGenerator.ts
@@ -1,26 +1,13 @@
 import { AbsoluteFilePath, join, RelativeFilePath, relative } from "@fern-api/fs-utils";
-import { createLoggingExecutable, runExeca } from "@fern-api/logging-execa";
+import { createLoggingExecutable } from "@fern-api/logging-execa";
 import { TaskContext } from "@fern-api/task-context";
-import { access, cp, readFile, unlink, writeFile } from "fs/promises";
+import { cp, readFile, unlink, writeFile } from "fs/promises";
 import tmp from "tmp-promise";
-import { getProtobufYamlV1 } from "./utils";
+import { detectAirGappedMode, getProtobufYamlV1 } from "./utils";
 
 const PROTOBUF_GENERATOR_CONFIG_FILENAME = "buf.gen.yaml";
 const PROTOBUF_GENERATOR_OUTPUT_PATH = "output";
 const PROTOBUF_GENERATOR_OUTPUT_FILEPATH = `${PROTOBUF_GENERATOR_OUTPUT_PATH}/openapi.yaml`;
-
-// Network error detection helper
-function isNetworkError(errorMessage: string): boolean {
-    return (
-        errorMessage.includes("server hosted at that remote is unavailable") ||
-        errorMessage.includes("failed to connect") ||
-        errorMessage.includes("network") ||
-        errorMessage.includes("ENOTFOUND") ||
-        errorMessage.includes("ETIMEDOUT") ||
-        errorMessage.includes("TIMEDOUT") ||
-        errorMessage.includes("timed out")
-    );
-}
 
 export class ProtobufOpenAPIGenerator {
     private context: TaskContext;
@@ -72,7 +59,7 @@ export class ProtobufOpenAPIGenerator {
     }): Promise<{ absoluteFilepath: AbsoluteFilePath; bufLockContents: string | undefined }> {
         // Detect air-gapped mode once at the start if we have dependencies
         if (deps.length > 0 && this.isAirGapped === undefined) {
-            await this.detectAirGappedMode(absoluteFilepathToProtobufRoot);
+            this.isAirGapped = await detectAirGappedMode(absoluteFilepathToProtobufRoot, this.context.logger);
         }
 
         const protobufGeneratorConfigPath = await this.setupProtobufGeneratorConfig({
@@ -86,76 +73,6 @@ export class ProtobufOpenAPIGenerator {
             deps,
             existingBufLockContents
         });
-    }
-
-    /**
-     * Detect if we're in an air-gapped environment by trying buf dep update once.
-     * This sets the isAirGapped flag which is used by subsequent operations.
-     */
-    private async detectAirGappedMode(absoluteFilepathToProtobufRoot: AbsoluteFilePath): Promise<void> {
-        const bufLockPath = join(absoluteFilepathToProtobufRoot, RelativeFilePath.of("buf.lock"));
-
-        // Check if buf.lock exists (required for air-gapped mode)
-        let bufLockExists = false;
-        try {
-            await access(bufLockPath);
-            bufLockExists = true;
-            this.context.logger.debug(`Found buf.lock at: ${bufLockPath}`);
-        } catch {
-            this.context.logger.debug(`No buf.lock found at: ${bufLockPath}`);
-        }
-
-        if (!bufLockExists) {
-            // No buf.lock means we need network access - not air-gapped
-            this.isAirGapped = false;
-            return;
-        }
-
-        // Try a quick network check with buf dep update using a short timeout
-        const tmpDir = AbsoluteFilePath.of((await tmp.dir()).path);
-        try {
-            // Copy buf.yaml and buf.lock to temp dir for the test
-            const bufYamlPath = join(absoluteFilepathToProtobufRoot, RelativeFilePath.of("buf.yaml"));
-            try {
-                await cp(bufYamlPath, join(tmpDir, RelativeFilePath.of("buf.yaml")));
-                await cp(bufLockPath, join(tmpDir, RelativeFilePath.of("buf.lock")));
-            } catch {
-                // If we can't copy the files, assume not air-gapped
-                this.isAirGapped = false;
-                return;
-            }
-
-            // Try buf dep update with a short timeout (5 seconds)
-            try {
-                await runExeca(this.context.logger, "buf", ["dep", "update"], {
-                    cwd: tmpDir,
-                    stdio: "pipe",
-                    timeout: 5000 // 5 second timeout for quick detection
-                });
-                // Network works - not air-gapped
-                this.isAirGapped = false;
-                this.context.logger.debug("Network check succeeded - not in air-gapped mode");
-            } catch (error) {
-                const errorMessage = error instanceof Error ? error.message : String(error);
-                if (isNetworkError(errorMessage)) {
-                    this.isAirGapped = true;
-                    this.context.logger.debug(
-                        `Network check failed - entering air-gapped mode: ${errorMessage.substring(0, 100)}`
-                    );
-                } else {
-                    // Non-network error - assume not air-gapped
-                    this.isAirGapped = false;
-                }
-            }
-        } finally {
-            // Cleanup temp dir
-            try {
-                await unlink(join(tmpDir, RelativeFilePath.of("buf.yaml")));
-                await unlink(join(tmpDir, RelativeFilePath.of("buf.lock")));
-            } catch {
-                // Ignore cleanup errors
-            }
-        }
     }
 
     private async setupProtobufGeneratorConfig({
@@ -232,50 +149,15 @@ export class ProtobufOpenAPIGenerator {
                 // If we're in air-gapped mode, skip buf dep update entirely
                 if (this.isAirGapped) {
                     this.context.logger.debug("Air-gapped mode: skipping buf dep update");
-                    // Read existing buf.lock contents for caching
-                    try {
-                        bufLockContents = await readFile(bufLockPath, "utf-8");
-                    } catch {
-                        bufLockContents = undefined;
-                    }
                 } else {
-                    // Try buf dep update to populate the cache (needed at build time)
-                    // Backup: If it fails with a network error and buf.lock exists, continue
-                    try {
-                        await buf(["dep", "update"]);
-                        // buf dep update succeeded, read the updated buf.lock
-                        try {
-                            bufLockContents = await readFile(bufLockPath, "utf-8");
-                        } catch {
-                            bufLockContents = undefined;
-                        }
-                    } catch (bufDepUpdateError: unknown) {
-                        const errorMessage =
-                            bufDepUpdateError instanceof Error ? bufDepUpdateError.message : String(bufDepUpdateError);
-
-                        // Check if buf.lock exists for backup fallback
-                        let bufLockExistsInCopiedDir = false;
-                        try {
-                            await access(bufLockPath);
-                            bufLockExistsInCopiedDir = true;
-                            bufLockContents = await readFile(bufLockPath, "utf-8");
-                        } catch {
-                            bufLockExistsInCopiedDir = false;
-                        }
-
-                        this.context.logger.debug(
-                            `buf dep update failed. isNetworkError=${isNetworkError(errorMessage)}, bufLockExists=${bufLockExistsInCopiedDir}, error=${errorMessage.substring(0, 200)}`
-                        );
-
-                        if (isNetworkError(errorMessage) && bufLockExistsInCopiedDir) {
-                            // Backup: Air-gapped environment with pre-cached buf.lock - continue
-                            this.context.logger.debug(
-                                "buf dep update failed due to network error, but buf.lock exists. Continuing (backup)."
-                            );
-                        } else {
-                            this.context.failAndThrow(errorMessage);
-                        }
-                    }
+                    // Run buf dep update to populate the cache (needed at build time)
+                    await buf(["dep", "update"]);
+                }
+                // Read buf.lock contents for caching
+                try {
+                    bufLockContents = await readFile(bufLockPath, "utf-8");
+                } catch {
+                    bufLockContents = undefined;
                 }
             }
